Allow add domain without protocol
closes #92
